+++ src/core/converter/ImportFileConverter.ts
@@ -3,7 +3,19 @@ import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 export abstract class ImportFileConverter {
     // static readonly fileExtension: string;
 
-    abstract convert(content: string, flightplan: AeroflyFlight): void;
+    /**
+     * In a given file there may be multiple flight plans present.
+     * This method is supposed to return the name as well as indices of the found flight plans.
+     * In most files there will be only a single flight plan included, so this will return a single string called "default".
+     */
+    getIndices(content: string): string[] {
+        if (content === "") {
+            throw new Error("No content for importing found");
+        }
+        return ["default"];
+    }
+
+    abstract convert(content: string, flightplan: AeroflyFlight, index: number): void;
 
     /**
      * This function is a placeholder until the method to encode UIDs is discovered.