+++ types/core/converter/ImportFileAeroflyCustomMissionsTmcConverter.d.ts
@@ -2,6 +2,7 @@ import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 export declare class ImportFileAeroflyCustomMissionsTmcConverter extends ImportFileConverter {
     static readonly fileExtension = "tmc";
-    convert(content: string, flightplan: AeroflyFlight): void;
+    getIndices(content: string): string[];
+    convert(content: string, flightplan: AeroflyFlight, index?: number): void;
 }
 //# sourceMappingURL=ImportFileAeroflyCustomMissionsTmcConverter.d.ts.map