+++ types/core/converter/ImportFileConverter.d.ts
@@ -1,6 +1,12 @@
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 export declare abstract class ImportFileConverter {
-    abstract convert(content: string, flightplan: AeroflyFlight): void;
+    /**
+     * In a given file there may be multiple flight plans present.
+     * This method is supposed to return the name as well as indices of the found flight plans.
+     * In most files there will be only a single flight plan included, so this will return a single string called "default".
+     */
+    getIndices(content: string): string[];
+    abstract convert(content: string, flightplan: AeroflyFlight, index: number): void;
     /**
      * This function is a placeholder until the method to encode UIDs is discovered.
      *