+++ types/core/converter/ImportMetarConverter.d.ts
@@ -1,6 +1,6 @@
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 import { ImportFileConverter } from "./ImportFileConverter.js";
 export declare class ImportMetarConverter extends ImportFileConverter {
-    convert(content: string, flightplan: AeroflyFlight): void;
+    convert(content: string, flightplan: AeroflyFlight, index?: number): void;
 }
 //# sourceMappingURL=ImportMetarConverter.d.ts.map