+++ types/core/io/ImportFileReader.d.ts
@@ -5,6 +5,7 @@ import { ImportFileConverter } from "../converter/ImportFileConverter.js";
  * appropriate converter class.
  */
 export declare class ImportFileReader {
+    static getFlightplans(filename: string): string[];
     /**
      * Imports a flight plan from a file and converts it to an AeroflyFlight object.
      * Supported file types are determined by the file extension:
@@ -13,12 +14,13 @@ export declare class ImportFileReader {
      *
      * @param filename The path to the flight plan file to import.
      * @param flightplan The AeroflyFlight object to populate with the imported data.
+     * @param index If multiple flight plans are present in a given file, select which index to import
      * @throws Will throw an error if the file type is unsupported or if the conversion fails.
      * @see ImportFileConverter for the interface that specific file handlers must implement.
      * @see ImportFileMsfs for handling Microsoft Flight Simulator .pln files.
-     * @see ImportFileGarminFpl for handling Garmin .fpl files.
+     * @see ImportFileGarminFplConverter for handling Garmin .fpl files.
      */
-    static importFile(filename: string, flightplan: AeroflyFlight): void;
+    static importFile(filename: string, flightplan: AeroflyFlight, index?: number): void;
     static getConverter(filename: string): new () => ImportFileConverter;
 }
 //# sourceMappingURL=ImportFileReader.d.ts.map