+++ types/core/util/SunPosition.d.ts
@@ -2,7 +2,12 @@ export declare class SunPosition {
     /**
      * Simplified elevation/ azimuth calculation
      */
-    static getSunPosition(utcHours: number, dayOfYear: number, latitude: number, longitude: number): {
+    static getSunPosition(
+        utcHours: number,
+        dayOfYear: number,
+        latitude: number,
+        longitude: number,
+    ): {
         elevation: number;
         azimuth: number;
     };